requirements/base.txt
django==4.0.3
django-ckeditor==6.2.0
django-compressor==3.1
django-debug-toolbar==3.2.4
django-extensions==3.1.5
django-logentry-admin==1.1.0
django-registration-redux==2.10
git+https://github.com/Doca/django-ajax.git#egg=djangoajax

python3-memcached==1.51

ipython==8.2.0
